+++ src/pagination/mod.rs
@@ -3,7 +3,7 @@ mod links;
 mod params;
 
 /// Maximum permitted page size to avoid expensive queries.
-const MAX_PER_PAGE: u32 = 500;
+const MAX_PER_PAGE: usize = 500;
 
 pub use info::PaginationInfo;
 pub use links::PageLink;
